+++ lib/gql/node-notification.d.ts
@@ -1,5 +1,4 @@
 import { Node } from './node.interface';
-import { Model } from 'sequelize';
 import { ClassType } from '../helpers/index';
 export declare enum NotificationOf {
     LAST_KNOWN_STATE = "LAST_KNOWN_STATE",
@@ -19,9 +18,3 @@ export declare abstract class NodeNotification<T extends Node<T>> {
     constructor(notificationOf: NotificationOf, node: T);
 }
 export declare function GqlNodeNotification<T extends Node<T>>(clsNotification: ClassType<T>): ClassType<NodeNotification<T>>;
-export declare class DbModelChangeNotification {
-    notificationOf: NotificationOf;
-    model: Model<any, any>;
-    deltas: ModelDelta[];
-    constructor(notificationOf: NotificationOf, model: Model<any, any>, deltas: ModelDelta[]);
-}