+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './gql';
 export * from './helpers';
 export * from './db';
 export * from './server';
+export * from './pubsub';